feat: add screenshot + refactor
