+++ package.json
@@ -4,7 +4,7 @@
   "description": "Template for the WebSocket remote control task",
   "main": "index.ts",
   "scripts": {
-    "start": "node index.ts",
+    "start": "nodemon ./index.ts",
     "start:dev": "nodemon ./index.ts",
     "test": "echo \"Error: no test specified\" && exit 1"
   },
@@ -20,6 +20,7 @@
   "dependencies": {
     "@nut-tree/nut-js": "^2.3.0",
     "@types/node": "^18.11.18",
+    "jimp": "^0.16.2",
     "ts-loader": "^9.4.2",
     "ws": "^8.12.0"
   }
