+++ src/drawing/rectangular.ts
