+++ src/drawing/square.ts
