+++ src/utils/parseData.ts
@@ -3,6 +3,5 @@ export const parseData = (command: string) => {
   const direction = array[0];
   const position = array[1];
   const figureLength = array[2];
-  // console.log(direction, position);
   return [direction, position, figureLength];
 };
